+++ src/Schedule/Schedule.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Solital\Core\Schedule;
+
+/**
+ * This class is used only for compatibility to Solital framework 4
+ */
+class Schedule extends TaskSchedule {}