+++ tests/Resource/SessionTest.php
@@ -1,5 +1,7 @@
 <?php
 
+require_once dirname(__DIR__) . '/bootstrap.php';
+
 use PHPUnit\Framework\TestCase;
 use Solital\Core\Resource\Session;
 
@@ -11,6 +13,10 @@ class SessionTest extends TestCase
     {
         Session::set('name', 'Solital Framework');
         $res = Session::get('name');
+        $this->assertEquals($res, 'Solital Framework');
+
+        session('name', 'Solital Framework');
+        $res = session('name');
         $this->assertEquals($res, 'Solital Framework');
     }
 
@@ -24,12 +30,21 @@ class SessionTest extends TestCase
     {
         $res = Session::delete('name');
         $this->assertTrue($res);
+
+        $res = session('name', delete: true);
+        $this->assertTrue($res);
     }
 
     public function testTakeSession()
     {
         Session::set('email', 'solital@email.com');
         $res = Session::take('email', 'message test');
+        $exists = Session::has('email');
+        $this->assertEquals($res, 'solital@email.com');
+        $this->assertFalse($exists);
+
+        session('email', 'solital@email.com');
+        $res = session('email', 'message test', take: true);
         $exists = Session::has('email');
         $this->assertEquals($res, 'solital@email.com');
         $this->assertFalse($exists);